fixed breaking json on closing editor, little code changes

+++ addons/gd-achievements/scripts/achievement_manager.gd
@@ -30,10 +30,6 @@ var soundNode = AudioStreamPlayer.new()
 # Main call signal that accepts index of an achievement from array
 signal showAchievement(index)
 
-func _init():
-	achievementsDataScript = load(ACHIEVEMENT_DATA_SCRIPT_PATH).new()
-	pass
-
 # Create audio node for playing sound
 func initSoundNode():
 	soundNode.set_stream(globalSound)
@@ -55,23 +51,23 @@ func rewriteAchievementsDataToUserJson():
 
 func updateReferenceJson(achievementsFromUserFileBuf):
 	if (globalAchievements.hash() != achievementsFromUserFileBuf.hash()):
-			print("AchievementSystem: User file and reference file are different. Updating data...")
+			print("AchievementSystem: User's file and reference file are different. Updating data...")
 			globalAchievements = achievementsFromUserFileBuf
 			print("AchievementSystem: Data updated!")
 			pass
 	pass
 
-func updateUserJson():
-	print("AchievementSystem: Updating data to JSON on device...")
+func updateJsonOnDeviceBeforeClosing():
 	var userFile = File.new()
-	userFile.open(ACHIEVEMENT_JSON_USER_PATH, File.WRITE)
-	userFile.store_string(to_json(globalAchievements))
-	userFile.close()
-	print("AchievementSystem: Updating completed!")
+	if (userFile.file_exists(ACHIEVEMENT_JSON_USER_PATH) and globalAchievements != null):
+		userFile.open(ACHIEVEMENT_JSON_USER_PATH, File.WRITE)
+		userFile.store_string(to_json(globalAchievements))
+		userFile.close()
+	print("AchievementSystem: updateJsonOnDeviceBeforeClosing()")
 	pass
 
-func checkFileJsonOnDevice():
-	print("AchievementSystem: Start checking 'achievements.json' file on device...")
+func checkJsonOnDevice():
+	print("AchievementSystem: Checking 'achievements.json' file on device...")
 	var jsonUserFile = File.new()
 	if (not jsonUserFile.file_exists(ACHIEVEMENT_JSON_USER_PATH)):
 		print("AchievementSystem: 'achievements.json' not found on device. Creating...")
@@ -81,8 +77,9 @@ func checkFileJsonOnDevice():
 		print("AchievementSystem: File created!")
 		pass
 	else:
-		print("AchievementSystem: File is exists!")
+		print("AchievementSystem:'achievements.json' is exists on device!")
 		jsonUserFile.open(ACHIEVEMENT_JSON_USER_PATH, File.READ)
+		
 		print("AchievementSystem: Getting data from file...")
 		var achievementsFromUserFileBuf = parse_json(jsonUserFile.get_as_text())
 		jsonUserFile.close()
@@ -98,17 +95,10 @@ func checkFileJsonOnDevice():
 
 func _notification(what):
 	if (what == MainLoop.NOTIFICATION_WM_QUIT_REQUEST):
-		updateUserJson()
+		updateJsonOnDeviceBeforeClosing()
 		pass
 	pass
 
-func _ready():
-	globalAchievements = achievementsDataScript.getAchievements()
-	initSoundNode()
-	checkFileJsonOnDevice()
-	connect("showAchievement", self, "activateAchievement")
-	pass
-	
 func getFieldName(index):
 	return(str(globalAchievements.keys()[index]))
 	pass
@@ -145,59 +135,74 @@ func resetAchievementNotifications():
 	for i in range((len(globalAchievements.keys()))):
 		globalAchievements.values()[i]["is_have"] = 0
 		pass
+	rewriteAchievementsDataToUserJson()
 	print("AchievementSystem: resetAchievementsNotifications()")
 	pass
 
+func _ready():
+	achievementsDataScript = load(ACHIEVEMENT_DATA_SCRIPT_PATH).new()
+	globalAchievements = achievementsDataScript.getAchievements()
+	initSoundNode()
+	checkJsonOnDevice()
+	connect("showAchievement", self, "activateAchievement")
+	pass
+	
+
+func createAchievementPanel(achievementIndex):
+	# Preload instance of scene
+	achievementUiNotificationInstance = preload(ACHIEVEMENT_UI_NOTIFICATION_PATH).instance()
+	# Get ID for a new instance
+	var achievementUiNotificationId = achievementUiNotificationInstance.get_instance_id()
+	# Create instance with unique ID for independent working with its own nodes
+	var achievementUiNotification = instance_from_id(achievementUiNotificationId)
+	add_child(achievementUiNotification)
+		
+	# Increase count of UI instances for correct position
+	achievementCount += 1
+
+	# Change position of current instance's animation if it more than 1 notification window
+	if (achievementCount > 1):
+		achievementUiNotification.rect_position.y = achievementCurrentPosY + 130
+		achievementCurrentPosY = achievementUiNotification.rect_position.y
+		pass
+		
+	# Set name
+	achievementUiNotification.get_node("achievementPanel/achievementDescription/description").text = globalAchievements.keys()[int(achievementIndex)]
+		
+	# Set icon path
+	achievementUiNotification.get_node("achievementPanel/achievementIcon/TextureRect").texture = load(
+		globalAchievements.values()[int(achievementIndex)]['icon_path']
+		)
+	# Play sound
+	soundNode.play()
+		
+	# Play animation
+	achievementUiNotification.get_node("AnimationPlayer").play("popup")
+		
+	# Wait a few seconds
+	yield(achievementUiNotification.get_tree().create_timer(SHOW_TIME), "timeout")
+	achievementCount -= 1
+		
+	# Play hide animation
+	achievementUiNotification.get_node("AnimationPlayer").play("hide")
+		
+	# Delete node
+	yield(get_tree().create_timer(SHOW_END_TIME), "timeout")
+	achievementUiNotification.queue_free()	
+	pass
+
 func activateAchievement(achievementIndex):
-	if (achievementIndex > len(globalAchievements.keys()) - 1):
+	if (achievementIndex > (len(globalAchievements.keys()) - 1) or achievementIndex < 0):
 		print("AchievementSystem Error: Attempt to get an achievement on " + achievementIndex + 
 			" index that is out of range (" + (len(globalAchievements.keys()) - 1) + ")")
 		return
+	
+	var currentAchievementIsHaveField = int(globalAchievements.values()[achievementIndex]["is_have"])
 
-	if ((achievementCount >= 0) and (achievementCount <= len(globalAchievements.keys()) - 1) and (globalAchievements.values()[achievementIndex]["is_have"] == 0)):
+	if ((achievementCount >= 0) and (achievementCount <= len(globalAchievements.keys()) - 1) and currentAchievementIsHaveField == 0):
 		print("AchievementSystem: Show achievement '" + str(globalAchievements.keys()[achievementIndex]) + "', index = " + str(achievementIndex))
 		globalAchievements.values()[achievementIndex]["is_have"] = 1
 		rewriteAchievementsDataToUserJson()
 		
-		# Preload instance of scene
-		achievementUiNotificationInstance = preload(ACHIEVEMENT_UI_NOTIFICATION_PATH).instance()
-		# Get ID for a new instance
-		var achievementUiNotificationId = achievementUiNotificationInstance.get_instance_id()
-		# Create instance with unique ID for independent working with its own nodes
-		var achievementUiNotification = instance_from_id(achievementUiNotificationId)
-		add_child(achievementUiNotification)
-		
-		# Increase count of UI instances for correct position
-		achievementCount += 1
-
-		# Change position of current instance's animation if it more than 1 notification window
-		if (achievementCount > 1):
-			achievementUiNotification.rect_position.y = achievementCurrentPosY + 130
-			achievementCurrentPosY = achievementUiNotification.rect_position.y
-			pass
-		
-		# Set name
-		achievementUiNotification.get_node("achievementPanel/achievementDescription/description").text = globalAchievements.keys()[int(achievementIndex)]
-		
-		# Set icon path
-		achievementUiNotification.get_node("achievementPanel/achievementIcon/TextureRect").texture = load(
-			globalAchievements.values()[int(achievementIndex)]['icon_path']
-			)
-		
-		# Play sound
-		soundNode.play()
-		
-		# Play animation
-		achievementUiNotification.get_node("AnimationPlayer").play("popup")
-		
-		# Wait a few seconds
-		yield(achievementUiNotification.get_tree().create_timer(SHOW_TIME), "timeout")
-		achievementCount -= 1
-		
-		# Play hide animation
-		achievementUiNotification.get_node("AnimationPlayer").play("hide")
-		
-		# Delete node
-		yield(get_tree().create_timer(SHOW_END_TIME), "timeout")
-		achievementUiNotification.queue_free()	
+		createAchievementPanel(achievementIndex)		
 	pass
